Started migrating some classes.
Some of these classes I am bringing over by simply importing the HTML da
ta into Markdown files.  This is ugly, but it is quite a bit of work to
preserve the formatting by reverse-engineering the files back from HTML
to Markdown.  I am using Pandoc to do this, but it loses some of the ess
ential HTML formatting.  So for now I'll keep the old ugly HTML and use
Markdown for the new classes.  One nice feature of the HTML parser inclu
ded in Nesta is that it is very picky; so at least the HTML is nice and
clean.