Updated to use Kramdown as the parser for Markdown.
Graham Ashton has updated Nesta so that the Markdown parser is
configurable.  This commit updates the site to use Kramdown, for two
reasons.  First, it is the only parser which implements in a clean way
the footnote syntax from PHP Markdown Extra.  As posted to the Nesta
mailing list, here is a summary of the different Markdown parsers with
respect to this feature:    - BlueCloth: Unsupported, as best I can
tell.  - Kramdown: Supported. See:
http://kramdown.rubyforge.org/syntax.html - Maruku: Unsupported, though
a fix is described here: https://github.com/nex3/maruku/issues/18 -
RDiscount: Unsupported, though there is a pull request adding the
feature here: https://github.com/rtomayko/rdiscount/pull/34 - Redcarpet:
Unsupported  Second, Kramdown has built-in support for math blocks,
which (like footnotes, come to mention it) is a very attractive feature
for academics.  Moreover, the lovely MathJax display engine is used to
render the math.  All in all, this is a very nice upgrade over Maruku,
the main branch of which hasn't been updated in over a year.