Now deployed on Railway.
Heroku removed their free plan, so the site has been migrated to Railway. Two recent changes to the site have been made to facilitate this transition:

1. The Procfile has been deleted.
2. A Dockerfile has been added.

This is because of the slightly different architecture used by Railway. Deploying to Railway simply requires committing to Github; Railway detects this change, and then automatically builds and redeploys the site.