Fix kramdown processing
Not sure when this problem was introduced, but old blog posts were not using the kramdown parser for markdown, and so the footnotes were not displaying correctly. The main update is to ensure, through an edit to the site theme, that kramdown is used for markdown in md, mdown and haml files.